API+DOC Move proximity to static.py Add msd, pair_correlation_2d and pair_correlation_3d to docs
