+++ doc/whatsnew.rst
@@ -7,4 +7,4 @@ What's new in the package
 
 A catalog of new features, improvements, and bug-fixes in each release. Follow links to the relevant GitHub issue or pull request for specific code changes and any related discussion.
 
-.. include:: releases/v0.3.0.txt
+.. include:: releases/v0.3.1.txt
